:goal_net: ♻️ string_handling.py - simpler error handling
In short, this way if the optional parameter `response_key` for the error message has been left blank, then there will simply be a double space in the error message.
It isn't the MOST elegant fix, but it saves a lot of boilerplate. PLUS, now I think it is easier to read.